+++ tests/test_02_dxf_graphics/test_226_spline.py
@@ -180,12 +180,13 @@ def test_set_uniform(spline, points):
     assert spline.closed is False
 
 
-def test_set_periodic(spline, points):
-    spline.set_periodic(points, degree=3)
-    assert spline.dxf.degree == 3
-    assert list(spline.control_points) == points
-    assert spline.dxf.n_control_points == len(points)
-    assert spline.dxf.n_knots == len(points) + 1  # according the Autodesk developer documentation
+def test_set_closed(spline, points):
+    # closed spline: first `degree` control points will be repeated at the end
+    degree = 3
+    spline.set_closed(points, degree=degree)
+    assert spline.dxf.degree == degree
+    assert spline.dxf.n_control_points == len(points) + degree
+    assert spline.dxf.n_knots == len(spline.control_points) + degree + 1  # n + p + 2
     assert spline.closed is True
 
 
@@ -211,14 +212,14 @@ def test_set_uniform_rational(spline, points, weights):
     assert spline.closed is False
 
 
-def test_set_periodic_rational(spline, points, weights):
-    spline.set_periodic_rational(points, weights, degree=3)
+def test_set_closed_rational(spline, points, weights):
+    # closed spline: first `degree` control points and weights will be repeated at the end
+    degree = 3
+    spline.set_closed_rational(points, weights, degree=degree)
     assert spline.dxf.degree == 3
-    assert list(spline.control_points) == points
-    assert list(spline.weights) == weights
-    assert spline.dxf.n_control_points == len(points)
-    assert len(spline.weights) == len(points)
-    assert spline.dxf.n_knots == len(points) + 1  # according the Autodesk developer documentation
+    assert spline.dxf.n_control_points == len(points) + degree
+    assert len(spline.weights) == len(points) + degree
+    assert spline.dxf.n_knots == len(spline.control_points) + degree + 1  # n + p + 2
     assert spline.closed is True
 
 
@@ -238,11 +239,12 @@ def test_open_spline(msp, points):
 
 
 def test_closed_spline(msp, points):
-    spline = msp.add_closed_spline(points, degree=3)
-    assert spline.dxf.degree == 3
-    assert list(spline.control_points) == points
-    assert spline.dxf.n_control_points == len(points)
-    assert spline.dxf.n_knots == len(points) + 1  # according the Autodesk developer documentation
+    degree = 3
+    spline = msp.add_closed_spline(points, degree=degree)
+    assert spline.dxf.degree == degree
+    assert list(spline.control_points) == points + points[:degree]
+    assert spline.dxf.n_control_points == len(spline.control_points)
+    assert spline.dxf.n_knots == len(spline.control_points) + degree + 1
     assert spline.closed is True
 
 
@@ -258,13 +260,14 @@ def test_open_rational_spline(msp, points, weights):
 
 
 def test_closed_rational_spline(msp, points, weights):
-    spline = msp.add_closed_rational_spline(points, weights, degree=3)
-    assert spline.dxf.degree == 3
-    assert list(spline.control_points) == points
-    assert list(spline.weights) == weights
-    assert spline.dxf.n_control_points == len(points)
-    assert len(spline.weights) == len(points)
-    assert spline.dxf.n_knots == len(points) + 1  # according the Autodesk developer documentation
+    degree = 3
+    spline = msp.add_closed_rational_spline(points, weights, degree=degree)
+    assert spline.dxf.degree == degree
+    assert list(spline.control_points) == points + points[:degree]
+    assert list(spline.weights) == weights + weights[:degree]
+    assert spline.dxf.n_control_points == len(points) + degree
+    assert len(spline.weights) == len(points) + degree
+    assert spline.dxf.n_knots == len(spline.control_points) + degree + 1
     assert spline.closed is True